Add logic to fetch show metadata when we add an item to the library.
- We will buildup on this to fetch and add upNext content.

+++ presenter/watchlist/src/commonMain/kotlin/com/thomaskioko/tvmaniac/presentation/watchlist/model/WatchlistItem.kt
@@ -5,8 +5,8 @@ data class WatchlistItem(
   val title: String,
   val posterImageUrl: String? = null,
   val watchProgress: Float = 0F,
-  val seasonCount: Int = 0,
-  val episodeCount: Int = 0,
+  val seasonCount: Long = 0,
+  val episodeCount: Long = 0,
   val status: String? = null,
   val year:String? = null
 )
